Hotfix/access permission (#7)
* fixing user management

* Fixing access permision and make dynamic menu

* adding middleware for permission in every route

+++ app/Http/Controllers/UserController.php
@@ -8,53 +8,51 @@ use Maatwebsite\Excel\Facades\Excel;
 use App\Imports\UsersImport;
 use Spatie\Permission\Models\Role;
 use Illuminate\Support\Facades\Auth;
-use App\Models\Ticket;
+use App\Models\User;
 
-
+use function App\Helpers\cek_akses_user;
+use function App\Helpers\main_menu;
+use function App\Helpers\sub_menu;
 
 class UserController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            if (Auth::check()) {
+                $this->cek = cek_akses_user();
+            }
+            //     // $this->sub_menu = sub_menu();
+            return $next($request);
+        });
+    }
+    
     //tampilkan data
     public function index(Request $request)
-    {
-        if(Auth::user()->roles[0]->name == "engineer"){
-            $count_open = Ticket::where("status", "open")->where("assign_id", null)->count();
-            $count_progress = Ticket::where("status", "progress")->where("assign_id", Auth::user()->id)->count();
-            $count_close = Ticket::where("status", "close")->where("assign_id", Auth::user()->id)->count();
-
-        }else if(Auth::user()->roles[0]->name == "kanim"){
-            $count_open = Ticket::where("status", "open")->where("assign_id", Auth::user()->id)->count();
-            $count_progress = Ticket::where("status", "progress")->where("assign_id", Auth::user()->id)->count();
-            $count_close = Ticket::where("status", "close")->where("assign_id", Auth::user()->id)->count();
-
-
-        }else {
-            $count_open = Ticket::where("status", "open")->count();
-            $count_progress = Ticket::where("status", "progress")->count();
-            $count_close = Ticket::where("status", "close")->count();
-        }
+    {        
         $users = \App\Models\User::paginate(5);
         $filterKeyword = $request->get('keyword');
            
         // $users = DB::table('users')->paginate(5);
         $roles = $request->get('roles');
         if($roles){
-            $users = \App\Models\User::where('roles', $roles)->paginate(5);
+            $users = User::role($roles)->paginate(5);
+            
+            // $users = \App\Models\User::where('roles', $roles)->paginate(5);
            } else {
             $users = \App\Models\User::paginate(5);
            }
            if($filterKeyword){
             if($roles){
-            $users = \App\Models\User::where('email', 'LIKE', "%$filterKeyword%")
-                ->where('roles', $roles)
-                ->paginate(5);
+            $users = User::role($roles)->where('email', 'LIKE', "%$filterKeyword%")->paginate(5);
             } else {
             $users = \App\Models\User::where('email', 'LIKE', "%$filterKeyword%")
                 ->paginate(5);
             }
-           }
-           
-        return view('users', compact("users", "count_open", "count_progress", "count_close"));
+        }
+        
+        return view('users.index', compact("users"));
+
     }
 
     //Method Validation
@@ -141,25 +139,13 @@ class UserController extends Controller
     //action untuk menampilkan form tambah data 
     public function tambah()
     {
-        if(Auth::user()->roles[0]->name == "engineer"){
-            $count_open = Ticket::where("status", "open")->where("assign_id", null)->count();
-            $count_progress = Ticket::where("status", "progress")->where("assign_id", Auth::user()->id)->count();
-            $count_close = Ticket::where("status", "close")->where("assign_id", Auth::user()->id)->count();
-
-        }else if(Auth::user()->roles[0]->name == "kanim"){
-            $count_open = Ticket::where("status", "open")->where("assign_id", Auth::user()->id)->count();
-            $count_progress = Ticket::where("status", "progress")->where("assign_id", Auth::user()->id)->count();
-            $count_close = Ticket::where("status", "close")->where("assign_id", Auth::user()->id)->count();
-
-
-        }else {
-            $count_open = Ticket::where("status", "open")->count();
-            $count_progress = Ticket::where("status", "progress")->count();
-            $count_close = Ticket::where("status", "close")->count();
+        if($this->cek->tambah != 1) {
+            abort(403);
         }
+        
         $kanims = \App\Models\Kanim::get();
         $roles = Role::get();
-        return view('users-tambah', compact("kanims", "roles", "count_open", "count_progress", "count_close"));
+        return view('users-tambah', compact("kanims", "roles"));
     }
 
     //method untuk simpan data 
@@ -190,31 +176,22 @@ class UserController extends Controller
     //method untuk edit data 
     public function edit($id)
     {
+        if($this->cek->edit != 1) {
+            abort(403);
+        }
         $user = \App\Models\User::findOrFail($id);
         $kanims = \App\Models\Kanim::get();
-        if(Auth::user()->roles[0]->name == "engineer"){
-            $count_open = Ticket::where("status", "open")->where("assign_id", null)->count();
-            $count_progress = Ticket::where("status", "progress")->where("assign_id", Auth::user()->id)->count();
-            $count_close = Ticket::where("status", "close")->where("assign_id", Auth::user()->id)->count();
-
-        }else if(Auth::user()->roles[0]->name == "kanim"){
-            $count_open = Ticket::where("status", "open")->where("assign_id", Auth::user()->id)->count();
-            $count_progress = Ticket::where("status", "progress")->where("assign_id", Auth::user()->id)->count();
-            $count_close = Ticket::where("status", "close")->where("assign_id", Auth::user()->id)->count();
-
-
-        }else {
-            $count_open = Ticket::where("status", "open")->count();
-            $count_progress = Ticket::where("status", "progress")->count();
-            $count_close = Ticket::where("status", "close")->count();
-        }
-        return view('users-edit', compact("user", "kanims", "count_open", "count_progress", "count_close"));
+        
+        return view('users-edit', compact("user", "kanims"));
     }
 
     //method untuk hapus data
 
     public function hapus($id)
     {
+        if($this->cek->hapus != 1) {
+            abort(403);
+        }
         // DB::table('users')->where('id',$id)->delete();
         $user = \App\Models\User::findOrFail($id);
         $user->delete();
@@ -225,6 +202,10 @@ class UserController extends Controller
 
     public function update(Request $request, $id)
     {
+        if($this->cek->edit != 1) {
+            abort(403);
+        }
+        //$this->_validation($request);
         $validation = \Validator::make($request->all(),[
             "name" => "required|min:5|max:100",
             // "username" => "required|min:5|max:20|",
@@ -256,7 +237,7 @@ class UserController extends Controller
 
         $user = \App\Models\User::findOrFail($id);
         $user->name = $request->get('name');
-        $user->roles = $request->get('roles');
+        $user->syncRoles($request->get('roles'));
         $user->phone = $request->get('phone');
         // $user->password = $request->get('password');
         if ($request->file('image')) {
@@ -275,22 +256,7 @@ class UserController extends Controller
     public function detail($id)
     {
         $user = \App\Models\User::findOrFail($id);
-        if(Auth::user()->roles[0]->name == "engineer"){
-            $count_open = Ticket::where("status", "open")->where("assign_id", null)->count();
-            $count_progress = Ticket::where("status", "progress")->where("assign_id", Auth::user()->id)->count();
-            $count_close = Ticket::where("status", "close")->where("assign_id", Auth::user()->id)->count();
-
-        }else if(Auth::user()->roles[0]->name == "kanim"){
-            $count_open = Ticket::where("status", "open")->where("assign_id", Auth::user()->id)->count();
-            $count_progress = Ticket::where("status", "progress")->where("assign_id", Auth::user()->id)->count();
-            $count_close = Ticket::where("status", "close")->where("assign_id", Auth::user()->id)->count();
-
-
-        }else {
-            $count_open = Ticket::where("status", "open")->count();
-            $count_progress = Ticket::where("status", "progress")->count();
-            $count_close = Ticket::where("status", "close")->count();
-        }
-        return view('users-detail',compact("user", "count_open", "count_progress", "count_close"));
+        
+        return view('users-detail',compact("user"));
     }
 }
